api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.1
alembic==1.13.1
pillow==10.2.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0